+++ lib/neuroimaging/fmri/fmristat/multistat.py
@@ -29,7 +29,7 @@ class RFXMean(traits.HasTraits):
 
         self.nsubject = len(input_files)
 
-        self.X = N.ones((self.nsubject, 1), N.Float)
+        self.X = N.ones((self.nsubject, 1), N.float64)
         self.Xsq = N.power(N.transpose(self.design_matrix), 2)
         self.pinvX = L.pinv(self.X)
 
@@ -64,7 +64,7 @@ class RFXMean(traits.HasTraits):
                         raise ValueError, 'len(df) != len(input_files) in MultiStat'
                     self.input_df = array(list(df))
                 except TypeError:
-                    self.input_df = N.ones((len(input_files),), N.Float) * df
+                    self.input_df = N.ones((len(input_files),), N.float64) * df
             else:
                 self.input_df = N.inf * N.ones((len(input_files),))
 
@@ -82,22 +82,22 @@ class RFXMean(traits.HasTraits):
 
     def estimate_varatio(self, Y, S=None, df=None):
 
-        self.Y = N.zeros((self.nsubject, self.npixel), N.Float)
-        self.S = N.ones((self.nsubject, self.npixel), N.Float)
+        self.Y = N.zeros((self.nsubject, self.npixel), N.float64)
+        self.S = N.ones((self.nsubject, self.npixel), N.float64)
 
     def fit(self, Y, S=None, df=None):
 
         if not self.fixed and self.varatio is None:
             self.estimate_varatio(Y, S, df)
             
-        effect = N.zeros(Y.shape[1:], N.Float)
-        sdeffect = N.zeros(Y.shape[1:], N.Float)
+        effect = N.zeros(Y.shape[1:], N.float64)
+        sdeffect = N.zeros(Y.shape[1:], N.float64)
 
         ncpinvX = N.sqrt(N.add.reduce(N.power(N.squeeze(self.pinvX), 2)))
    
         sigma2 = self.varfix * self.varatio
 
-        Sigma = self.S + multiply.outer(N.ones((self.nsubject,), N.Float), sigma2)
+        Sigma = self.S + multiply.outer(N.ones((self.nsubject,), N.float64), sigma2)
 
         # To ensure that Sigma is always positive:
         if self.fwhm_varatio > 0:
@@ -114,7 +114,7 @@ class RFXMean(traits.HasTraits):
         varatio_smooth.setshape(product(varatio_smooth.shape))
         sigma2 = N.transpose(varatio_smooth)
         sdeffect = ncpinvX * sqrt(N.transpose(sigma2))
-        Sigma = N.dot(N.one((self.nsubject,), N.Float), sigma2)
+        Sigma = N.dot(N.one((self.nsubject,), N.float64), sigma2)
         W = recipr(Sigma)
         effect = N.transpose(N.dot(self.contrast, betahat))
